Fix monitoring of monitoring logs
The script I use to check backups considers an empty file to be a failure.  Unfortunately there are no failures some days and so the monit logs are empty.  This appends something to the end of the file so that the check passes on good days as well as bad.

I've made the change in production and already checked that it works :)